check that fqdn's IP addr belongs to a net iface (SOFTWARE-3389)
this addresses the issue where the hostname is a lie.

Unfortunately we had to bring in (what seems like) a ton of code just
to get the list of net ifaces and their IP addrs.

As mentioned in the comments, credits for the net addr/ip list code,
lifted from:

https://github.com/opensciencegrid/htcondor-ce/blob/v3.1.3/src/condor_ce_host_network_check

Based on http://programmaticallyspeaking.com/getting-network-interfaces-in-python.html
which is, in turn, based on Based on getifaddrs.py from pydlnadms [http://code.google.com/p/pydlnadms/].