Update ucsb-hep/ucsb-cms contacts
Adam Dishaw (former grid admin at UCSB) is gone and said in an email
that Jae Hyeok Yoo took over his duties.  If this is not still current,
we'll fix that also.

In any case, we are removing refernces to Adam Dishaw, as we are
removing his contact from our contacts repo, since his email is no
longer valid and is producing bounces.