add github pull_request webhook handler to app.py (SOFTWARE-3325)
This is a squash of edquist:wip/SOFTWARE-3325.downtime-automerge-webhook,
which contains the original history of this webhook.

I'll reference one commit here below from the many squashed, as we may
want to go back to using `pull/N/merge` / `merge_commit_sha`, if we only
want to allow auto-merging the initial PR (action="opened"), rather than
on every push to the head branch (action="synchronize").

---

commit 547d418020fcf4dca30b9fd9915b9dd172b7a9a1
Author: Carl Edquist <edquist@cs.wisc.edu>
Date:   Tue Oct 16 18:38:19 2018 -0500

    use pull/N/head instead of merge_commit_sha (SOFTWARE-3325)

    apparently pull/N/merge and merge_commit_sha only get updated when the
    PR is initially created, and then upon explicit api request, which then
    requires further polling until the mergeable attribute is set ...

    instead we'll just do a separate check that the pull commit contains the
    base commit, and we now explicitly need to fetch both refs.

    https://developer.github.com/enterprise/2.1/changes/2013-04-25-deprecating-merge-commit-sha/

    https://developer.github.com/v3/git/#checking-mergeability-of-pull-requests