Revert the resource name changes
We already have existing payload records with a Host_description set
to the above resource names, generated by Gratia. Then through GRACC
summarizer magic, the Host_description is cross-referenced against the
topology data to get the site name.

So if we want to fix the site name in the GRACC dashboards, we just
need to a) register the resource so that it matches the
Host_description (OSG_US_NMSU_AGGIE_GRID) or b) change the site name
of the resource corresponding to the Host_description (OSG_US_NMSU_DISCOVERY)