Unscheduled CIT downtime
There are power issues in one of the racks and also several other datanodes are down and affect all HDFS. Files are not lost, but nodes are not accessible and this affects all SE. 

p.s. There is no xrootd as a service for SE, shouldn't this be implemented in some way, same as xrootd caches? Or this is not something what CMS would look/need?