fall back to merge-base for out-of-date head commits
Well, it's one way to do it, and it works even if there are
merge conflicts.

However, it's kind of broken in that the PR submitter can
pick an arbitrary merge base in the upstream base's history,
and while the list of files changed ought still to be accurate,
the version of the resource group files may differ, which means
the wrong stuff might still be getting validated.

Most likely it won't be a problem at all if the PR head was
relatively recent, and in any case the script will always
report failure since it was out-of-date; however, the main
concern is that a malicious submitter "fool" the email viewers
into thinking that the only issue is the PR head is out-of-date,
when in fact there will be other errors once updated.

Probably the github-provided merge commit should be used, if
available, and if not then fall back to this merge-base strategy