accept separate HEAD_SHA and (optional) MERGE_COMMIT_SHA
and move arg parsing to a separate function

the script expected a MERGE_COMMIT_SHA but is passed an unmerged
HEAD_SHA in the case where the PR head was not up-to-date.

making this distinction will allow us to check whether HEAD_SHA is
up-to-date, but fall back to MERGE_COMMIT_SHA, if it exists, for the
purpose of showing files changed.