add functions for oasis managers endpoint (SOFTWARE-3947)
For an authorized request, the new endpoint will take a VO and
return the associated OASIS Managers with ssh pubkeys and DNs.

The new `get_oasis_manager_endpoint_info` function facilitates
collecting and organizing all the relevant information from
contacts, topology, and cilogon ldap to return the information
for this endpoint.