behave yourself if CILOGON_LDAP_PASSFILE not specified (SOFTWARE-4528)
well, all three (CILOGON_LDAP_{URL|USER|PASSFILE}) are needed for it
to work, so if any of them are None, back out gracefully.  Just, the
PASSFILE one is the only one without a default, so it's the only
one that actually needs to be specified.

But yeah, even without it nothing should explode, which is what
this fixes.