supplement comanage data with contact db info (SOFTWARE-4734)
merge_yaml_data used just to add the contact db items to the
comanage contact data, as separate contacts; but we actually want
to add the missing info to the comanage contacts themselves.

per SOFTWARE-4734, we want to match on CILogonID, if possible, otherwise
match on email address.  Primary or Secondary email from either source
can match against either one in the other source.

i've broken the work for this into a handful of smaller functions in
order to help keep my sanity, though there is a growing feeling that
all of these, essentially dict manipulation functions could cozily
find themselves in a separate module, rather than with the proper
cilogon ldap functions.  i'm open to suggestions if anyone has a
strong feeling on that.