normalize emails to lowercase for merging contacts
As noticed while inspecting ldap results for SOFTWARE-5313, comanage
apparently sometimes stores email addresses in all-caps.
That ought to be OK, since cilogon_id_map_to_yaml_data() stores them
in lowercase.  But the email_lookup table generated from the contacts
db does not do this, and sometimes emails are actually stored in
mixed-case.  So we need to build the email_lookup table based on the
normalized lowercase in order to merge on email addresses properly.

This issue (bug) may have resulted in some contacts failing to merge
automatically on email address.