tweak for run within topology
the whitelist file is no longer under a 'checks' dir.  also while we are
at it we can get rid of some of the automatic logic in slurp to figure
out whether the file is local or remote -- since we run from topology
now instead of the cvmfs config repo, the configs are always remote.

make a separate function for slurping a file, which is used specifically
for the whitelist.