move TIGER-OSG-BACKFILL to CHTC prod
Per https://support.opensciencegrid.org/a/tickets/69078, this resource
is serving prod jobs and thus should be marked for Production.

I do not want to mark the whole CHTC-ITB resource group for Production,
so I am moving just the resource in question to the CHTC resource group
which is Production.