mark chameleon-uc-dev-backfill for Production
Per https://support.opensciencegrid.org/a/tickets/69078, this resource
is serving prod jobs and thus should be marked for Production.

There are no other resources in this resource group, so I am marking the
whole resource group as Production.  (There is no other prod resource
group equivalent at this Site.)  Alternatively, the resource group could
be renamed, but I do not want to mess with that.