Add HEPScore23Percentage to WLCGInformation
HEPScore23Percentage is the percentage of the resource that has been benchmarked
with HEPScore23 benchmark, the new benchmark for HEP.  Initially, it would be 0.  Sites
will set this to the percent of their resources benchmarked with the new benchmark.  
Since the old benchmark (HEPSPEC) and the new (HEPScore23) have a 1 to 1 normalization
factor, no changes to the APELNormalFactor are required.  How the percentage of a
resource is defined, it left as an exercise for the reader.  Percent of Cores?  Percent of
job slots?  Percent of Memory?