Add new LIGO OSDF Origins to CIT_LIGO.yaml
This PR adds two new OSDF origin servers to the Caltech LIGO resource group. I have marked them as inactive because while the servers are up and should respond to ping, we have not finished configuring XRootD on them.  However, if they need to be marked active in order for [this test](https://osg-htc.org/docs/data/stashcache/install-origin/#testing-file-access-authenticated-origin) to work, then please when considering this PR also change those to active. I have not added ID lines for the servers themselves as my understanding is that those will be assigned to them as part of the PR approval.