Add tests for stash/osdf scitokens.conf generation
This also adds a test VO and ResourceGroup that's only visible to the tests,
and test code to load them.  This will allow us to test generation on fake
data and not have the code break due to production data changes.

This was originally written for PR #3457 but we changed our minds about that
feature.