Update SPRACE_downtime.yaml
In early February there was an electrical outage in our campus, and due to circumstances beyond our control we were forced to shutdown the whole data center. Unfortunately that operation affected several services provided by our data center, including the equipment that securely holds the root certificate for our Certification Authority (ANSPGrid CA), which is used to sign the grid certificates for all SPRACE servers. The new equipment was configured but a new root certficate needed to be created and all certificates needs to be re-issued.