Add downtime for BNL-ATLAS due to dCache maintenance
This will involve work on the hardware supporting the core database systems, as well as a database upgrade, and will require the system to be stopped during the intervention.
 
In addition, we will take the opportunity to migrate residual dCache components currently hosted in RHEV, and to update the dCache minor version as needed. This also provides a chance to enable monitoring data collectors that require a dCache cell restart.