Change capitalization of Hyunwoo's name to match contacts DB
The contacts.yaml entry for Hyunwoo has his name in all caps (and that's how it's used in the resources he's a contact for).
This entry has been causing CI failures in the contact repo, so change it to match.

(I could have also changed the contact.yaml entry and his name in all the resources, but this is lower impact.)